% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixer_functions.R
\name{fix_time}
\alias{fix_time}
\title{Fixes the time column of a data set}
\usage{
fix_time(df)
}
\arguments{
\item{df}{a table with a time column from Statistics Denmark. Usually retrieved
using the \link{retrieve_data} function.}
}
\value{
a data frame
}
\description{
Statistics Denmark data often come with times/dates that are formatted to
fit months or quarters. This function converts them in to proper date strings
in the format YYYY-MM-DD.
}
\examples{
df <- retrieve_data("BEV3A", BEVÆGELSEV = "B02", Tid = "2017M06,2017M05")
dplyr::glimpse(df)

df <- fix_time(df)
dplyr::glimpse(df)

}
