% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_functions.R
\name{retrieve_subjects}
\alias{retrieve_subjects}
\title{Retrieves an overview of the subjects from Statistics Denmark}
\usage{
retrieve_subjects(base_url = "http://api.statbank.dk/v1/")
}
\arguments{
\item{base_url}{is the base url for the API you wish to call. Statistics Denmark
can sometimes create custom API's that you can use by changing this
parameter.}
}
\value{
a tibble
}
\description{
The function retrieves a data frame with the over all statistical subjects
that can be pulled from Statistics Denmark.
}
\examples{
subjects <- statsDK::retrieve_subjects()
dplyr::glimpse(subjects)
}
